Avoid clone for sparse tensors during accumulation of grads.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33427

This PR is an attempt to avoid clone for sparse tensors similar to how
we avoid clone for dense tensors currently.

As per my understanding even if the 'indices' and 'values' of a sparse tensor
are non-continguous, operations like 'add' are still supported. As a result,
the major change in this PR is to use create a shallow copy instead of clone()
for sparse tensors.
ghstack-source-id: 98999537

Differential Revision: [D19926698](https://our.internmc.facebook.com/intern/diff/D19926698/)